# Use a C array for transparency LUT for debug perf