--- a/Source/engine/palette.h
+++ b/Source/engine/palette.h
@@ -36,7 +36,7 @@ extern std::array<SDL_Color, 256> logical_palette;
 extern std::array<SDL_Color, 256> system_palette;
 extern std::array<SDL_Color, 256> orig_palette;
 /** Lookup table for transparency */
-extern std::array<std::array<Uint8, 256>, 256> paletteTransparencyLookup;
+extern Uint8 paletteTransparencyLookup[256][256];
 
 #if DEVILUTIONX_PALETTE_TRANSPARENCY_BLACK_16_LUT
 /**