In a debug build, `std::array` accesses are function calls. This mostly does not matter but this particular array is read from a lot on every frame.

Timedemo on my machine:
* Before: 45.39 seconds	141.3 FPS
* After:  40.20 seconds	159.5 FPS